Update analysis; Major refactor
The bulk of the previous work of prodrop_analysis now happens in
ProdropAnalyzer class. This was done to lessen the pain of dealing with
so much data being tracked and so many counters, but it makes for nice
decoupling as well.

A ReportWriter class was also written, and is used in writing the
analysis reports.